@@ -14,16 +14,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.4",
+    "@angular/animations": "^9.1.2",
     "@angular/cdk": "^8.1.4",
-    "@angular/common": "^8.2.4",
-    "@angular/compiler": "^8.2.4",
-    "@angular/core": "^8.2.4",
-    "@angular/forms": "^8.2.4",
+    "@angular/common": "^9.1.2",
+    "@angular/compiler": "^9.1.2",
+    "@angular/core": "^9.1.2",
+    "@angular/forms": "^9.1.2",
     "@angular/material": "^8.1.4",
-    "@angular/platform-browser": "^8.2.4",
-    "@angular/platform-browser-dynamic": "^8.2.4",
-    "@angular/router": "^8.2.4",
+    "@angular/platform-browser": "^9.1.2",
+    "@angular/platform-browser-dynamic": "^9.1.2",
+    "@angular/router": "^9.1.2",
     "@aspnet/signalr": "^1.1.4",
     "@types/bootstrap": "^3.3.33",
     "@types/jquery": "^3.3.31",
@@ -66,17 +66,18 @@
     "sweetalert": "^2.0.8",
     "toastr": "^2.1.2",
     "ts-helpers": "^1.1.2",
+    "tslib": "^1.10.0",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.2",
-    "@angular/cli": "^8.3.2",
-    "@angular/compiler-cli": "^8.2.4",
+    "@angular-devkit/build-angular": "~0.901.2",
+    "@angular/cli": "^9.1.2",
+    "@angular/compiler-cli": "^9.1.2",
     "@angularclass/hmr": "^2.1.3",
     "@types/jasmine": "^3.4.0",
-    "@types/node": "^12.7.3",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "^3.3.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^4.3.0",
@@ -89,7 +90,7 @@
     "protractor": "^5.1.1",
     "ts-node": "^8.3.0",
     "tslint": "^5.19.0",
-    "typescript": "3.4.5",
+    "typescript": "3.8.3",
     "webpack-bundle-analyzer": "^3.0.3"
   }
 }
